Split notifications
