+++ src/Clock/View.elm
@@ -19,6 +19,7 @@ import PomodoroCss
 { class, id } =
   Html.CssHelpers.withNamespace ""
 
+styles : List Css.Mixin -> Html.Attribute msg
 styles =
   Css.asPairs >> Html.Attributes.style
